Add String Editor Function, Add a bunch of function
MyImGui.cpp - Make a list of helper function, Completely implement the function that print out an information of string. Form an outline of the function that highlight the selected object. have to be Implemented later.
Object.cpp&hpp - Make a function that return a string of the object type.
String.cpp&hpp - return a const reference of vertices in string. In order to editor represent the information 
of string.

CIB - Robin